Included the lightweight rendering pipeline, Core RP library and lightweight RP packages so that a shader graph can be used to create the hologram shader.
Updated all the materials to use the lightweight rendering pipeline since standard materials did not work with it.

In the XR settings, changed the stereo rendering mode to Single Pass instead of multi pass. Apparently, the lightweight rendering pipeline uses Single Pass rendering
which caused the headset to only render on one eye when it was in multi pass mode.

A first pass of the hologram effect we want is implemented.